requirements.txt
pyquery==1.4.0
PyVirtualDisplay==0.2.1
pymongo==3.6.0
psutil==5.4.3
requests==2.18.4
PyMySQL==0.9.3
Pillow==5.3.0
